cuRandom refactor
