+++ lib/random.cu
@@ -44,7 +44,7 @@ namespace quda {
       int idd
         = (((x[3] * arg.commDim[2] * arg.X[2] + x[2]) * arg.commDim[1] * arg.X[1]) + x[1]) * arg.commDim[0] * arg.X[0]
         + x[0];
-      curand_init(seed, idd, 0, &state[parity * size_cb + id]);
+      random_init(seed, idd, 0, &state[parity * size_cb + id]); 
     }
   }
